Fixed problem with pumping beam squiggle not appearing until the beam control was moved.

+++ trunk/lasers/src/edu/colorado/phet/lasers/model/photon/CollimatedBeam.java
@@ -102,7 +102,7 @@ public class CollimatedBeam extends Particle {
         eventRegistry.fireEvent( new RateChangeEvent() );
     }
 
-    public void setWavelength( int wavelength ) {
+    public void setWavelength( double wavelength ) {
         this.wavelength = wavelength;
         WavelengthChangeEvent event = new WavelengthChangeEvent();
         eventRegistry.fireEvent( event );
